refactor: melhora performance

+++ representacao_leitura/matriz_adjacencia.h
@@ -3,13 +3,14 @@
 
 #include <vector>
 #include <string>
+#include <cstdint>
 #include "leitor_grafo.h"
 using namespace std;
 
 class MatrizAdjacencia {
 private:
     int n; // número de vértices
-    vector<vector<int>> matriz;
+    vector<vector<uint8_t>> matriz;  // uint8_t economiza 75% da memória (4→1 byte)
 
 public:
     // Construtor que recebe dados do grafo
@@ -22,7 +23,7 @@ public:
     void imprimir() const;
     
     // Getter para acessar a matriz (útil para BFS)
-    const vector<vector<int>>& getMatriz() const;
+    const vector<vector<uint8_t>>& getMatriz() const;
 };
 
 #endif
